crawl/city_intro/requirements.txt
scrapy
selenium
cassandra-driver